:construction::truck: box/cluster/controller: Offload control plane implementation to _controlside.py module.
